25: added role verification for credit cards

+++ server/controllers/creditCardController.js
@@ -1,5 +1,7 @@
 /* eslint-disable no-underscore-dangle */
+const xssFilters = require('xss-filters');
 const CreditCard = require('../models/CreditCard');
+const UserRole = require('../models/UserRole');
 
 module.exports = {
   /**
@@ -8,41 +10,45 @@ module.exports = {
    * @param {Object} res is express response object
    */
   async listCreditCards(req, res) {
-    try {
-      const cards = await CreditCard.find();
-      const currentURL = `${req.protocol}://${req.get('host')}${
-        req.originalUrl
-      }`;
-      const cardsWithLinks = [];
-      cards.forEach((card) => {
-        const currentCard = JSON.parse(JSON.stringify(card));
-        // Add HATEOAS links to credit card
-        currentCard.links = [
-          {
-            rel: 'self',
-            method: 'GET',
-            href: `${currentURL}/${currentCard._id}`,
-            types: ['application/json'],
-          },
-          {
-            rel: 'self',
-            method: 'DELETE',
-            href: `${currentURL}/${currentCard._id}`,
-            types: [],
-          },
-          {
-            rel: 'self',
-            method: 'PUT',
-            href: `${currentURL}/${currentCard._id}`,
-            types: ['application/json'],
-          },
-        ];
-        cardsWithLinks.push(currentCard);
-      });
-      res.status(200).json(cardsWithLinks);
-    } catch (err) {
-      res.status(404).json({ message: err });
-      console.log('Caught an error: ', err);
+    if (req.userRole === UserRole.ADMIN) {
+      try {
+        const cards = await CreditCard.find();
+        const currentURL = `${req.protocol}://${req.get('host')}${
+          req.originalUrl
+        }`;
+        const cardsWithLinks = [];
+        cards.forEach((card) => {
+          const currentCard = JSON.parse(JSON.stringify(card));
+          // Add HATEOAS links to credit card
+          currentCard.links = [
+            {
+              rel: 'self',
+              method: 'GET',
+              href: `${currentURL}/${currentCard._id}`,
+              types: ['application/json'],
+            },
+            {
+              rel: 'self',
+              method: 'DELETE',
+              href: `${currentURL}/${currentCard._id}`,
+              types: [],
+            },
+            {
+              rel: 'self',
+              method: 'PUT',
+              href: `${currentURL}/${currentCard._id}`,
+              types: ['application/json'],
+            },
+          ];
+          cardsWithLinks.push(currentCard);
+        });
+        res.status(200).json(cardsWithLinks);
+      } catch (err) {
+        res.status(404).json({ message: err });
+        console.log('Caught an error: ', err);
+      }
+    } else {
+      res.status(401).json({ message: 'Unauthorized' });
     }
   },
 
@@ -53,38 +59,42 @@ module.exports = {
    */
 
   async showCreditCard(req, res) {
-    try {
-      const currentURL = `${req.protocol}://${req.get('host')}${
-        req.originalUrl
-      }`;
-      let card = await CreditCard.findById(req.params.cardId).exec();
-      card = JSON.parse(JSON.stringify(card));
+    if (req.userRole) {
+      try {
+        const currentURL = `${req.protocol}://${req.get('host')}${
+          req.originalUrl
+        }`;
+        let card = await CreditCard.findById(req.params.cardId).exec();
+        card = JSON.parse(JSON.stringify(card));
 
-      // Add HATEOAS links to credit card
-      card.links = [
-        {
-          rel: 'self',
-          method: 'GET',
-          href: currentURL,
-          types: ['application/json'],
-        },
-        {
-          rel: 'self',
-          method: 'DELETE',
-          href: currentURL,
-          types: [],
-        },
-        {
-          rel: 'self',
-          method: 'PUT',
-          href: currentURL,
-          types: ['application/json'],
-        },
-      ];
-      res.status(200).json(card);
-    } catch (err) {
-      res.status(404).json({ message: err });
-      console.log('Caught an error: ', err);
+        // Add HATEOAS links to credit card
+        card.links = [
+          {
+            rel: 'self',
+            method: 'GET',
+            href: currentURL,
+            types: ['application/json'],
+          },
+          {
+            rel: 'self',
+            method: 'DELETE',
+            href: currentURL,
+            types: [],
+          },
+          {
+            rel: 'self',
+            method: 'PUT',
+            href: currentURL,
+            types: ['application/json'],
+          },
+        ];
+        res.status(200).json(card);
+      } catch (err) {
+        res.status(404).json({ message: err });
+        console.log('Caught an error: ', err);
+      }
+    } else {
+      res.status(401).json({ message: 'Unauthorized' });
     }
   },
 
@@ -94,30 +104,34 @@ module.exports = {
    * @param {Object} res is express response object
    */
   async addCreditCard(req, res) {
-    try {
-      const { number, CVC, ownerName } = req.body;
-      let card = await CreditCard.findOne({ number }).exec();
+    if (req.userRole) {
+      try {
+        const { number, CVC, ownerName } = req.body;
+        let card = await CreditCard.findOne({ number }).exec();
 
-      if (card) {
-        const errorMessage = 'Card already registered in DB!';
+        if (card) {
+          const errorMessage = 'Card already registered in DB!';
 
-        if (!req.is('json')) {
-          res.status(400).json({ message: errorMessage });
+          if (!req.is('json')) {
+            res.status(400).json({ message: errorMessage });
+          }
         }
+
+        card = new CreditCard({
+          number: xssFilters.inHTMLData(number),
+          CVC: xssFilters.inHTMLData(CVC),
+          ownerName: xssFilters.inHTMLData(ownerName),
+        });
+
+        const newCard = await card.save();
+        await res.status(201).json(newCard);
+        console.log('New credit card added.');
+      } catch (err) {
+        res.json({ message: err });
+        console.log('Caught an error: ', err);
       }
-
-      card = new CreditCard({
-        number,
-        CVC,
-        ownerName,
-      });
-
-      const newCard = await card.save();
-      await res.status(201).json(newCard);
-      console.log('New credit card added.');
-    } catch (err) {
-      res.json({ message: err });
-      console.log('Caught an error: ', err);
+    } else {
+      res.status(401).json({ message: 'Unauthorized' });
     }
   },
 
@@ -127,14 +141,18 @@ module.exports = {
    * @param {Object} res is express response object
    */
   async removeCreditCard(req, res) {
-    try {
-      const ID = req.params.cardId;
-      const deletedCard = await CreditCard.deleteOne({ _id: `${ID}` });
-      res.status(200).json(deletedCard);
-      console.log('Successfully removed credit card.');
-    } catch (err) {
-      res.status(400).json({ message: err });
-      console.log('Caught an error: ', err);
+    if (req.userRole) {
+      try {
+        const ID = req.params.cardId;
+        const deletedCard = await CreditCard.deleteOne({ _id: `${ID}` });
+        res.status(200).json(deletedCard);
+        console.log('Successfully removed credit card.');
+      } catch (err) {
+        res.status(400).json({ message: err });
+        console.log('Caught an error: ', err);
+      }
+    } else {
+      res.status(401).json({ message: 'Unauthorized' });
     }
   },
 
@@ -144,17 +162,21 @@ module.exports = {
    * @param {Object} res is express response object
    */
   async updateCreditCard(req, res) {
-    try {
-      const ID = req.params.cardId;
-      const { number, CVC, ownerName } = req.body;
-      const updatedCard = await CreditCard.findOne({ _id: `${ID}` });
-      updatedCard.number = number;
-      updatedCard.CVC = CVC;
-      updatedCard.ownerName = ownerName;
-      await updatedCard.save();
-      res.status(200).json(updatedCard);
-    } catch (err) {
-      res.status(400).json({ message: err });
+    if (req.userRole) {
+      try {
+        const ID = req.params.cardId;
+        const { number, CVC, ownerName } = req.body;
+        const updatedCard = await CreditCard.findOne({ _id: `${ID}` });
+        updatedCard.number = xssFilters.inHTMLData(number);
+        updatedCard.CVC = xssFilters.inHTMLData(CVC);
+        updatedCard.ownerName = xssFilters.inHTMLData(ownerName);
+        await updatedCard.save();
+        res.status(200).json(updatedCard);
+      } catch (err) {
+        res.status(400).json({ message: err });
+      }
+    } else {
+      res.status(401).json({ message: 'Unauthorized' });
     }
   },
 };
